test-requirements.txt
django==5.1.3
django-admin-sortable2==2.2.4
django-bootstrap5==24.3
django-filebrowser-no-grappelli==4.0.2
django-lite-cms-core==0.0.7
django-mptt==0.16.0
django-mptt-admin==2.7.0
django-tinymce==4.1.0
Pillow==11.0.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
ruff==0.8.0
coverage==7.6.7
Faker==33.0.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0